Fix layouting issue with longer labels
On https://anmeldung.slub-dresden.de/en/ we have the label “Accept library rules and regulations *“ which is too long and pushes _only_ the required indicator onto the next line.

Since the right hand side of the form has a width of 60% we still have some slight wiggle room to compensate for this. If the while line breaks that’s fine. But at least on a regular desktop it should show as a single line.